+++ revision-v3.1/settings.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 export H728_INPUT_DIR=/armbian/cache/h728-v3.1-input
-export H728_CACHE_DIR=/armbian/cache/h728-v3.1.1
+export H728_CACHE_DIR=/armbian/cache/h728-v3.1.3
 export H728_COMMON=$H728_INPUT_DIR/base-common.sh
-export H728_IMAGE=$H728_CACHE_DIR/v3.1.1-working.img
-export H728_KERNEL_PACKAGE=/armbian/output/debs/linux-image-h728-manjaro_7.2.0-7+h728.5_arm64.deb
-export H728_KERNEL_PACKAGE_VERSION=7.2.0-7+h728.5
+export H728_IMAGE=$H728_CACHE_DIR/v3.1.3-working.img
+export H728_KERNEL_PACKAGE=/armbian/output/debs/linux-image-h728-manjaro_7.2.0-7+h728.7_arm64.deb
+export H728_KERNEL_PACKAGE_VERSION=7.2.0-7+h728.7
 export H728_WIFI_MAX_FREQUENCY=24000000